Create HTML template and game script
Add basic HTML template with head, title, body and script. Add the
computerPlay function, the playerPlay function, the playRound
function and the game function.

The HTML template is the minimun requirement to create an
interface for playing the game. The computerPlay function emulates
the a second player action, the playerPlay function collects the
player input and garantees that it is a valid one. The playRound
function executes a round of the game and the game function is
responsible for all the game by creating a player score, playing
5 rounds and displaying the final result.