Create the HTML layout of the page
Change title, add link to stylesheet and create fontawesome setup.
Add a header with title and subtitle. Add a scoreboard with title,
player score and computer score. Add buttons for rock, paper and
scissors. Add a footer with author's name and link to GitHub.

These elements where added to set the basic layout of the page for
Javascript to act on and to follow the project requirements.