Implement displays and winner announcer
Add a display showing the results of each round. Implement the
scoreboard, now it shows the players points. Add a window alert
announcing the winner and requesting a refresh to replay the game.
Rework some of the logic used in the script.

These additions are not only requirements of the project but also
used to improve the user interface since it removes the use of the
console.