Add style to the page
Change the subtitle and the font family. Add some colors to the
text and buttons. Align the score and buttons. Add a hover and
active animation for the buttons. Create an overflowing div with
the right size for the page. Adapt the page for responsiveness.

Although these changes were not required by the project, they
improve the layout of the page and embellish it. Besides, the
responsiveness is essential to mantain the improvement and
embellishment to all medias.